+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 0.1.2
+
+- Expose internal stats: `runningCount`, `waitingCount`, `scheduledCount`.
+- Enable sync-points with `join()`, which allows to track the completion of
+  the currently running (and optionally the waiting) tasks.
+
 ## 0.1.1
 
 - Fix missing heartbeat check on `Executor.close()`, without it the `Future`